requirements.txt
tensorflow==2.16.2
django==5.1.6
joblib==1.4.2
scikit-learn==1.3.1
numpy==1.26.4
django-cors-headers
python-dateutil